+++ minix/servers/vfs/filedes.c
@@ -508,7 +508,8 @@ close_filp(struct filp * f, int may_suspend)
 	f->filp_count = 0;
   } else if (f->filp_count < 0) {
 	panic("VFS: invalid filp count: %d ino %llx/%llu", f->filp_count,
-	      vp->v_dev, vp->v_inode_nr);
+	      (unsigned long long)vp->v_dev,
+	      (unsigned long long)vp->v_inode_nr);
   } else {
 	unlock_vnode(f->filp_vno);
   }
