+++ minix/servers/vfs/mount.c
@@ -179,7 +179,8 @@ char mount_label[LABEL_MAX] )
 	/* Get driver process' endpoint */
 	dp = &dmap[major(dev)];
 	if (dp->dmap_driver == NONE) {
-		printf("VFS: no driver for dev %llx\n", dev);
+		printf("VFS: no driver for dev %llx\n",
+		    (unsigned long long)dev);
 		return(EINVAL);
 	}
 
@@ -479,7 +480,8 @@ int unmount(
   /* Find vmnt that is to be unmounted */
   for (vmp_i = &vmnt[0]; vmp_i < &vmnt[NR_MNTS]; ++vmp_i) {
 	  if (vmp_i->m_dev == dev) {
-		  if(vmp) panic("device mounted more than once: %llx", dev);
+		  if(vmp) panic("device mounted more than once: %llx",
+		      (unsigned long long)dev);
 		  vmp = vmp_i;
 	  }
   }
@@ -579,7 +581,7 @@ void unmount_all(int force)
   for (vmp = &vmnt[0]; vmp < &vmnt[NR_MNTS]; vmp++) {
 	if (vmp->m_dev != NO_DEV) {
 		panic("vmp still mounted: %s %d %llx\n", vmp->m_label,
-			vmp->m_fs_e, vmp->m_dev);
+		    vmp->m_fs_e, (unsigned long long)vmp->m_dev);
 	}
   }
 }
