+++ minix/servers/vfs/open.c
@@ -268,7 +268,8 @@ int common_open(char path[PATH_MAX], int oflags, mode_t omode, int for_exec)
 			break;
 		   default:
 			printf("VFS: attempt to open file <%llu,%llu> of "
-			    "type 0%o\n", vp->v_dev, vp->v_inode_nr,
+			    "type 0%o\n", (unsigned long long)vp->v_dev,
+			    (unsigned long long)vp->v_inode_nr,
 			    vp->v_mode & S_IFMT);
 			r = EIO;
 		}
